tests/wpt/web-platform-tests/css/css-fonts/parsing/font-family-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 3: parsing font-family with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-3/#font-family-prop">
<meta name="assert" content="font-family supports the full grammar '[ <family-name> | <generic-family> ] #'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value('font-family', 'Serif', 'serif');
test_valid_value('font-family', 'Sans-Serif', 'sans-serif');
test_valid_value('font-family', 'Cursive', 'cursive');
test_valid_value('font-family', 'Fantasy', 'fantasy');
test_valid_value('font-family', 'Monospace', 'monospace');
test_valid_value('font-family', 'serif, sans-serif, cursive, fantasy, monospace');

test_valid_value('font-family', 'Helvetica, Verdana, sans-serif');
test_valid_value('font-family', '"New Century Schoolbook", serif');
test_valid_value('font-family', "'21st Century', fantasy", '"21st Century", fantasy');

test_valid_value('font-family', '"inherit", "serif"');
</script>
</body>
</html>
